add data modules for home page components
Removed the data needed for rendering from the component modules in
the home-page dir and moved them to their own modules. This makes the
component a little cleaner while keeping encapsulation entact.
