+++ src/components/home-page/choose-us-section/Choose-us.jsx
@@ -1,30 +1,7 @@
 import React from 'react';
 import PerkBox from './PerkBox';
-import coffeeBean from '../../../assets/icons/icon-coffee-bean.svg';
-import gift from '../../../assets/icons/icon-gift.svg';
-import truck from '../../../assets/icons/icon-truck.svg';
+import perks from './data';
 import styles from './_choose-us.module.scss';
-
-const perks = [
-  {
-    perkTitle: 'Best Quality',
-    perkDescription:
-      "Discover an endless variety of the world's artisan coffee from each of our roasters",
-    icon: coffeeBean,
-  },
-  {
-    perkTitle: 'Exclusive benefits',
-    perkDescription:
-      'Special offers and swag when you subscribe, including 30% off your first shipment.',
-    icon: gift,
-  },
-  {
-    perkTitle: 'Free shipping',
-    perkDescription:
-      '  We cover the cost and coffee is delivered fast. Peak freshness: guaranteed.',
-    icon: truck,
-  },
-];
 
 const ChooseUs = () => {
   return (
