+++ src/components/home-page/plan-section/Plan.jsx
@@ -1,25 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import StepBox from './StepBox';
+import steps from './data';
 import styles from './_plan.module.scss';
-
-const steps = [
-  {
-    title: 'Pick your coffee',
-    description:
-      'Select from our evolving range of artisan coffees. Our beans are ethically sourced and we pay fair prices for them. There are new coffees in all profiles every month for you to try out.',
-  },
-  {
-    title: 'Choose the frequency',
-    description:
-      'Customize your order frequency, quantity, even your roast style and grind type. Pause, skip or cancel your subscription with no commitment through our online portal.',
-  },
-  {
-    title: 'Receive and enjoy!',
-    description:
-      ' We ship your product within 48 hours, freshly roasted. Sit back and enjoy award-winning world-class coffees curated to provide a distinct tasting experience.',
-  },
-];
 
 const Plan = () => {
   return (
